+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+v2.7.0
+----------
+ * Add history message retrieval to client registration
+ * Enhance websocket payload handling to support 'message_with_fields' type
+
+v2.6.0
+----------
+ * Add CTAMessage support to Message struct and related tests
+
+v2.5.0
+----------
+ * Implement JWT authentication in all flows client requests
+
+v2.4.0
+----------
+ * Add support for updating contact fields
+
+ v2.3.0
+----------
+ * New build image
+
+v2.2.0
+----------
+ * APM basic integration
+
+v2.1.1
+----------
+ * Fix: add support for list messages
+
+v2.1.0
+----------
+ * Feat: improve redis memory management
+
 v2.0.4
 ----------
  * Enhance error handling in websocket client by adding benign error codes for non-standard frames